refactor: Update JobCatData import in Page component
